Add global disable to frames
This is important for times when entirely disabling the framing
mechanisms is needed. An example scenario is when building complex
queries for reporting that need full and unmodified access to the
tables in the database. These types of reports are typically explicitly
constructed for a certain result and framing could have unintended
consequences to those queries.